// Frontend/Angular/src/app/clients/teacher/model/userListDto.ts
/**
 * Teacher api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserType } from './userType';

export interface UserListDto { 
    userId?: number;
    name?: string;
    surname?: string;
    middlename?: string;
    userType?: UserType;
    photoId?: number;
    thumbnail?: string;
}